Use defineModel
